# faster structarray creation