Much faster for small arrays:
```julia
julia> @btime StructArray(a=1:1, b=[2], c=[:a])
  111.975 ns (4 allocations: 192 bytes)  # before
  56.601 ns (2 allocations: 128 bytes)   # after

julia> @btime StructArray(a=1:1, b=2:2, c=3.:3)
  87.747 ns (3 allocations: 208 bytes)   # before
  1.750 ns (0 allocations: 0 bytes)      # after
```